Revert "Shorten the protocol sequence field to 32 bits"
This reverts commit 0170bf3a8eb7ace21aed5c9246fa1d3b761b6b27.

+++ client/comm.cc
@@ -67,7 +67,7 @@
 #include "client_core.h"
 #include "comm.h"
 
-uint32_t Comm::sequence = 0L;
+uint64_t Comm::sequence = 0LL;
 
 /* Jump table for protocol handling */
 Comm::pkt_handler Comm::pkt_type[] =
